+++ ChangeLog.md
@@ -2,3 +2,4 @@
 
 ## Unreleased changes
 - Render parent Pom hierarchy for given project using Graphviz
+- Identify property overrides